refactor: 関数名の変更
