+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Container } from "@chakra-ui/react";
 import React, { useState } from "react";
-import Main from "./components/countdown/Main";
+import MainCountdown from "./components/countdown/MainCountdown";
 
 function App() {
   const time = new Date();
@@ -9,7 +9,11 @@ function App() {
 
   return (
     <Container maxW="container.sm">
-      <Main settingTime={settingTime} setSettingTime={setSettingTime} expiryTimestamp={time} />
+      <MainCountdown
+        settingTime={settingTime}
+        setSettingTime={setSettingTime}
+        expiryTimestamp={time}
+      />
     </Container>
   );
 }
